Create Assignment1_Q1sol
Assignment1
Q1. Write a program called Fibonacci to display the first 20 Fibonacci numbers F(n), where F(n)=F(n–1)+F(n–2) and F(1)=F(2)=1. Also compute their average. The output shall look like:

The first 20 Fibonacci numbers are:
1 1 2 3 5 8 13 21 34 55 89 144 233 377 610 987 1597 2584 4181 6765
The average is 885.5

Q2. Write a program called GradesAverage, which prompts user for the number of students, reads it from the keyboard, and saves it in an int variable called numStudents. It then prompts user for the grades of each of the students and saves them in an int array called grades. Your program shall check that the grade is between 0 and 100. A sample session is as follow:

Enter the number of students: 3 Enter the grade for student 1: 55 Enter the grade for student 2: 108 Invalid grade, try again...
Enter the grade for student 2: 56 Enter the grade for student 3: 57 The average is: 56.0

Q3. Write a method called copyOf(), which accept an int array and returns a copy of the given array. 
The method's signature is as follows:
public static int[] copyOf(int[] array) 

Q4. Create pascal triangle using 2D arrary
1
1 2
1 2 3
1 2 3 4
1 2 3 4 5
1 2 3 4 5 6
1 2 3 4 5 6 7
1 2 3 4 5 6 7 8

Q5. Write a program to find next date 
Input :  enter day: 31
 	 enter month: 12
	enter year : 2022
	today date: 31/12/2022
	Next date: 1/1/2023